Handle all events from one listener
The prior method depended on event bubbling behavior to properly set game state. But we already set the game state with handy variables like interstitialShowing. This method attaches one set of listeners on load and never changes them. Within the listener we check the game state and execute the appropriate click action.