Fix easier tapping
What we had before, although it was a copy of the balls array, was a copy of an array of objects. So all the objects were updating, meaning we weren't actually tracking the game state.

Instead we do collision detection and save the ball reference so that we can pop it on pointerup.

We also give the user the benefit of the doubt for popping - letting them pop either on pointerdown or pointerup, whichever they got the ball on.